+++ generation/main.go
@@ -22,14 +22,6 @@ func main() {
 	outputFile := flag.String("outputFile", "", "Output file")
 	chainID := flag.Uint64("chainID", 0, "Chain ID")
 
-	//WithdrawaProof specific flags
-	historicalSummariesIndex := flag.Uint64("historicalSummariesIndex", 0, "Historical summaries index")
-	blockHeaderIndex := flag.Uint64("blockHeaderIndex", 0, "Block header index")
-	historicalSummaryStateFile := flag.String("historicalSummaryStateFile", "", "Historical summary state file")
-	blockHeaderFile := flag.String("blockHeaderFile", "", "Block Header file")
-	blockBodyFile := flag.String("blockBodyFile", "", "Block Body file")
-	withdrawalIndex := flag.Uint64("withdrawalIndex", 0, "Withdrawal index")
-
 	// Parse the flags
 	flag.Parse()
 
@@ -44,9 +36,6 @@ func main() {
 	switch *command {
 	case "ValidatorFieldsProof":
 		err = GenerateValidatorFieldsProof(*oracleBlockHeaderFile, *stateFile, *validatorIndex, *chainID, *outputFile)
-
-	case "WithdrawalFieldsProof":
-		err = GenerateWithdrawalFieldsProof(*oracleBlockHeaderFile, *stateFile, *historicalSummaryStateFile, *blockHeaderFile, *blockBodyFile, *validatorIndex, *withdrawalIndex, *historicalSummariesIndex, *blockHeaderIndex, *chainID, *outputFile)
 
 	case "BalanceUpdateProof":
 		err = GenerateBalanceUpdateProof(*oracleBlockHeaderFile, *stateFile, *validatorIndex, *chainID, *outputFile)
