+++ merkle_util_test.go
@@ -164,96 +164,6 @@ func TestProveValidatorContainers(t *testing.T) {
 	assert.True(t, flag, "State Root Proof %v failed")
 }
 
-func TestProveWithdrawals(t *testing.T) {
-	oracleStateFile := "data/deneb_goerli_slot_7431952.json"
-	oracleHeaderFile := "data/deneb_goerli_block_header_7431952.json"
-	oracleStateJSON, err := ParseJSONFile(oracleStateFile)
-	if err != nil {
-		fmt.Println("error with JSON parsing beacon state")
-	}
-	oracleBlockHeader, err = ExtractBlockHeader(oracleHeaderFile)
-	if err != nil {
-		fmt.Println("error with block header", err)
-	}
-
-	ParseDenebBeaconStateFromJSON(*oracleStateJSON, &oracleState)
-
-	versionedOracleState, err := beacon.CreateVersionedState(&oracleState)
-	if err != nil {
-		fmt.Println("error creating versioned state", err)
-		return
-	}
-
-	historicalSummaryStateJSON, err := ParseJSONFile("data/deneb_goerli_slot_7421952.json")
-	if err != nil {
-		fmt.Println("error parsing historicalSummaryState JSON")
-	}
-	var historicalSummaryState deneb.BeaconState
-	ParseDenebBeaconStateFromJSON(*historicalSummaryStateJSON, &historicalSummaryState)
-	historicalSummaryStateBlockRoots := historicalSummaryState.BlockRoots
-
-	withdrawalBlock, err := ExtractBlockDeneb("data/deneb_goerli_block_7421951.json")
-	if err != nil {
-		fmt.Println("block.UnmarshalJSON error", err)
-	}
-	withdrawalBlockHeader, err := ExtractBlockHeader("data/deneb_goerli_block_header_7421951.json")
-	if err != nil {
-		fmt.Println("blockHeader.UnmarshalJSON error", err)
-	}
-
-	versionedWithdrawalBlock, err := beacon.CreateVersionedSignedBlock(withdrawalBlock)
-	if err != nil {
-		fmt.Println("error", err)
-		return
-	}
-
-	withdrawalIndex := uint64(0)
-	withdrawalValidatorIndex := uint64(627559) //this is the index of the validator with the first withdrawal in the withdrawalBlock 7421951
-
-	verifyAndProcessWithdrawalCallParams, err := epp.ProveWithdrawals(
-		&oracleBlockHeader,
-		&versionedOracleState,
-		[][]phase0.Root{historicalSummaryStateBlockRoots},
-		[]*spec.VersionedSignedBeaconBlock{&versionedWithdrawalBlock},
-		[]uint64{withdrawalValidatorIndex},
-	)
-	if err != nil {
-		fmt.Println("error", err)
-	}
-
-	executionPayloadRoot, err := withdrawalBlock.Body.ExecutionPayload.HashTreeRoot()
-	if err != nil {
-		fmt.Println("error", err)
-	}
-
-	oracleStateRoot, err := oracleState.HashTreeRoot()
-	if err != nil {
-		fmt.Println("error", err)
-		return
-	}
-
-	flag := verifyStateRootAgainstBlockHeaderProof(oracleBlockHeader, oracleState, verifyAndProcessWithdrawalCallParams.StateRootProof.StateRootProof)
-	assert.True(t, flag, "State Root Proof %v failed")
-
-	flag = verifyValidatorAgainstBeaconState(&oracleState, verifyAndProcessWithdrawalCallParams.ValidatorFieldsProofs[0], withdrawalValidatorIndex)
-	assert.True(t, flag, "Validator Fields Proof %v failed")
-
-	flag = verifyWithdrawalAgainstExecutionPayload(executionPayloadRoot, verifyAndProcessWithdrawalCallParams.WithdrawalProofs[0].WithdrawalProof, withdrawalIndex, withdrawalBlock.Body.ExecutionPayload.Withdrawals[0])
-	assert.True(t, flag, "Withdrawal Proof %v failed")
-
-	flag = verifyTimestampAgainstExecutionPayload(executionPayloadRoot, verifyAndProcessWithdrawalCallParams.WithdrawalProofs[0].TimestampProof, withdrawalBlock.Body.ExecutionPayload.Timestamp)
-	assert.True(t, flag, "Timestamp Proof %v failed")
-
-	flag = verifyBlockRootAgainstBeaconStateViaHistoricalSummaries(
-		oracleStateRoot,
-		verifyAndProcessWithdrawalCallParams.WithdrawalProofs[0].HistoricalSummaryBlockRootProof,
-		withdrawalBlockHeader,
-		verifyAndProcessWithdrawalCallParams.WithdrawalProofs[0].BlockRootIndex,
-		verifyAndProcessWithdrawalCallParams.WithdrawalProofs[0].HistoricalSummaryIndex,
-	)
-	assert.True(t, flag, "Historical Summary Block Root Proof %v failed")
-}
-
 func TestUnmarshalSSZVersionedBeaconStateDeneb(t *testing.T) {
 	oracleStateBytes, err := oracleState.MarshalSSZ()
 	if err != nil {
